refactor stats page
